// src/enums/SPGroupInternal.ts
export enum SPGroupInternal {
    UUO_RM = "Relationship manager",
    UUO_RSMA = "RSM Analyst",
    UUO_BO = "Back office",
    UUO_DEP = "UUO Deputy",
    UUO_DIR = "UUO Director",
    ORKR_RA = "Risk analyst",
    ORKR_GCCA = "GCC analyst",
    ORKR_RD = "Risk deputy",
    ORKR_RAPP = "Risk director",
    OAO_CC = "Credit controller",
    OAO_LAW = "Lawyer",
    OAO_DEP = "OAO Deputy",
    OAO_DIR = "OAO Director",
    WO_S = "WorkOut Specialist",
    WO_DEP = "WorkOut Deputy",
    WO_DIR = "WorkOut Director",
    AML_S = "AML Specialist",
    AML_DIR = "AML Director",
    APPR = "Approver",
    INT_AUD = "Interní Auditor",
    BUS_ADMIN = "Business Admin",
    ADMIN = "Admin"
}